+++ CHANGELOG.md
@@ -244,5 +244,6 @@
 * PlayMusic no longer restarts songs if they are already playing
 * Fixed NRE when reloading save on SpriteActorAnimationComponent
 * Added HasVideo API to VideoModule
+* Added ConditionalFilter that filters action special triggers by evaluating a Conditional
 * Fixed ThemeableElement and NonThemeableElement being ignored on child components of complex elements [Untested]
 * Fixed dialogue reactivating hidden objects even if they were initially inactive [Untested]